\name{declarativewidgets-package}
\alias{declarativewidgets-package}
\alias{declarativewidgets}
\docType{package}
\title{
\packageTitle{declarativewidgets}
}
\description{
\packageDescription{declarativewidgets}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{declarativewidgets}
\packageIndices{declarativewidgets}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{declarativewidgets}

Maintainer: \packageMaintainer{declarativewidgets}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
